Exercise 106
Paste content from exercise88.rkt and exercise92.rkt.
This provides functions for VCat and VCham.
A VAnimal is either:
- a VCat
- a VCham
So, the functions to
 - render
 - handle clock ticks
 - handle key envents
(a) ask whether the given VAnimal is a VCat or a VCham, and
(b) call the function that corresponds.

This differs from exercise 107 where the data definition has BOTH a VCat and a VCham.
For exercise 107, we make a new zoo world state that has updates to the field for the
VCat and updates to the field for VCham.