Exercise 107
Paste contents from exercise88.rkt and exercise92.rkt, so have VCat and VCham
and the functions for them: render, on-key, on-tick, stop-when.

(define-struct zoo [vcat vcham animal-in-focus])
A Zoo is a stucture:
  (make-zoo VCat VCham String)
interpretation:
   (make-zoo (make-cat-position-and-happiness x0 h0 d0)
             (make-virtual-chameleon-world-state x1 h1 c1)
             f0)
   represents a zoo with a cat and a chameleon. The cat
   has x-coordinate x0, happiness h0, and direction d0.
   The chameleon has x-coordinate x1, happiness h1, and
   color c1. The f0 animal is in focus.
   animal-in-focus is one of "cat" "chameleon"

render uses the function that renders a VCat and the function that renders the VCham.
We edit the function that renders a VCat so that the function takes as input
an image to put the VCat onto. Then we can compose the function for rendering
the VCat with the function to render the VCham.

key event can switch focus from one pet to the other.
key event otherwise checks which pet is in focus then calls the key handler for that type
e.g.
(make-zoo (key-handler (zoo-vcat zoo-world-state) key-event)
     	  (zoo-vcham zoo-world-state)
          (zoo-animal-in-focus zoo-world-state))

(make-zoo (zoo-vcat zoo-world-state)
          (update-due-to-key-event-virtual-chameleon-world-state (zoo-vcham zoo-world-state) key-event)
          (zoo-animal-in-focus zoo-world-state))

on-tick updates each field separately, i.e.
(make-zoo (update-cat-world-state (zoo-vcat zoo-world-state))
          (update-due-to-time-virtual-chameleon-world-state (zoo-vcham zoo-world-state))
          (zoo-animal-in-focus zoo-world-state))

stop-when checks whether we should stop because the VCat should stop or because the VCham should stop.